Renamed eraser.js to EraserTool.js
