+++ class/tools/EraserTool.js
